Update getting values
