+++ pkg/server/logging.go
@@ -10,20 +10,24 @@ import (
 func LogMW(logger *zap.SugaredLogger) Middleware {
 	return func(next http.Handler) http.Handler {
 		var h http.HandlerFunc = func(w http.ResponseWriter, r *http.Request) {
+			defer func() {
+				v := getValues(r)
+				if v == nil {
+					return
+				}
+
+				logger.Infow("request",
+					"request_id", v.TraceID,
+					"method", v.Method,
+					"path", v.RequestPath,
+					"status", v.StatusCode,
+					"duration", time.Since(v.Now),
+				)
+
+			}()
+
 			next.ServeHTTP(w, r)
 
-			v, ok := r.Context().Value(KeyValues).(*Values)
-			if !ok {
-				return
-			}
-
-			logger.Infow("request",
-				"request_id", v.TraceID,
-				"method", v.Method,
-				"path", v.RequestPath,
-				"status", v.StatusCode,
-				"duration", time.Since(v.Now),
-			)
 		}
 
 		return h
